feat: share the frameless title-bar script across platforms with macos native drag and a custom title bar docs page
